requirements.txt
fastapi==0.115.12
openai==1.69.0
pydantic==2.11.0
python-dotenv==1.1.0
uvicorn==0.34.0
pypdf
langchain_community